+++ frontend/src/app/main/data/common.cljs
@@ -7,7 +7,10 @@
 (ns app.main.data.common
   "A general purpose events."
   (:require
+   [app.config :as cf]
+   [app.main.data.messages :as msg]
    [app.main.repo :as rp]
+   [app.util.i18n :refer [tr]]
    [beicon.core :as rx]
    [potok.core :as ptk]))
 
@@ -43,3 +46,33 @@
     (watch [_ _ _]
       (->> (rp/cmd! :delete-share-link {:id id})
            (rx/ignore)))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; NOTIFICATIONS
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn force-reload!
+  []
+  (.reload js/location))
+
+(defn handle-notification
+  [{:keys [message code level] :as params}]
+  (ptk/reify ::show-notification
+    ptk/WatchEvent
+    (watch [_ _ _]
+      (case code
+        :upgrade-version
+        (when (or (not= (:version params) (:full cf/version))
+                  (true? (:force params)))
+          (rx/of (msg/dialog
+                  :content (tr "notifications.by-code.upgrade-version")
+                  :controls :inline-actions
+                  :type level
+                  :actions [{:label "Refresh" :callback force-reload!}]
+                  :tag :notification)))
+
+        (rx/of (msg/dialog
+                :content message
+                :controls :close
+                :type level
+                :tag :notification))))))